Update estimate_fee response
