+++ README.md
@@ -341,7 +341,7 @@ More detailed documentation can be found [here](#account).
 ```typescript
 it("should estimate fee", async function () {
     const fee = await contract.estimateFee("increase_balance", { amount: 10n });
-    console.log("Estimated fee:", fee.amount, fee.unit);
+    console.log("Estimated fee:", fee.amount, fee.unit, fee.gas_price, fee.gas_amount);
 });
 ```
